# Reduce SHA cgo calls